[Rename]: 파일경로 수정, 파일/폴더명 컨벤션에 맞게 수정
