+++ src/components/Main/MainSlide/MainSlide.js
