+++ src/components/Main/MainSlide/MainSlide.module.scss
@@ -63,7 +63,7 @@
       text-align: center;
       height: 55px;
       background-position: center center;
-      background-image: url('../../../public/image/mainSlide/nextIcon.png');
+      background-image: url('../../../../public/image/main/mainSlide/nextIcon.png');
       background-repeat: no-repeat;
       background-color: rgba(255, 255, 255, 0.253);
       background-size: 60%;
